+++ CHANGELOG.md
@@ -25,6 +25,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Spec files have more sanity checks and some keys are now optional.
 * Misc plom-manager improvements.
 * API calls have a default 10s timeout (and 2 retries), so semi-inevitable failures fail faster.
+* Do not mark pages in spec now given directly as list "doNotMarkPages = [1,2,3]"
+	
+	
 
 ### Fixed
 * Potential memory leaks in Annotator, Manager, and dialogs.